arch: rename xxxx_pause.c to xxxx_smpcall.c
