+++ arch/arm/src/armv7-a/CMakeLists.txt
@@ -109,7 +109,7 @@ if(CONFIG_ARCH_FPU)
 endif()
 
 if(CONFIG_SMP)
-  list(APPEND SRCS arm_cpustart.c arm_cpupause.c arm_cpuidlestack.c arm_scu.c)
+  list(APPEND SRCS arm_cpustart.c arm_smpcall.c arm_cpuidlestack.c arm_scu.c)
 endif()
 
 if(CONFIG_ARCH_HAVE_PSCI)
