+++ arch/arm/src/armv7-a/Make.defs
@@ -94,7 +94,7 @@ ifeq ($(CONFIG_ARCH_FPU),y)
 endif
 
 ifeq ($(CONFIG_SMP),y)
-  CMN_CSRCS += arm_cpustart.c arm_cpupause.c arm_cpuidlestack.c
+  CMN_CSRCS += arm_cpustart.c arm_smpcall.c arm_cpuidlestack.c
   CMN_CSRCS += arm_scu.c
 endif
