+++ arch/arm/src/armv7-a/arm_smpcall.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * arch/arm/src/armv7-a/arm_cpupause.c
+ * arch/arm/src/armv7-a/arm_smpcall.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -47,10 +47,10 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: arm_pause_async_handler
+ * Name: arm_smp_sched_handler
  *
  * Description:
- *   This is the handler for async pause.
+ *   This is the handler for sched.
  *
  *   1. It saves the current task state at the head of the current assigned
  *      task list.
@@ -66,7 +66,7 @@
  *
  ****************************************************************************/
 
-int arm_pause_async_handler(int irq, void *context, void *arg)
+int arm_smp_sched_handler(int irq, void *context, void *arg)
 {
   int cpu = this_cpu();
 
@@ -75,7 +75,7 @@ int arm_pause_async_handler(int irq, void *context, void *arg)
 }
 
 /****************************************************************************
- * Name: up_cpu_pause_async
+ * Name: up_send_smp_sched
  *
  * Description:
  *   pause task execution on the CPU
@@ -93,11 +93,30 @@ int arm_pause_async_handler(int irq, void *context, void *arg)
  *
  ****************************************************************************/
 
-inline_function int up_cpu_pause_async(int cpu)
+int up_send_smp_sched(int cpu)
 {
-  arm_cpu_sgi(GIC_SMP_CPUPAUSE_ASYNC, (1 << cpu));
+  arm_cpu_sgi(GIC_SMP_SCHED, (1 << cpu));
 
   return OK;
+}
+
+/****************************************************************************
+ * Name: up_send_smp_call
+ *
+ * Description:
+ *   Send smp call to target cpu.
+ *
+ * Input Parameters:
+ *   cpuset - The set of CPUs to receive the SGI.
+ *
+ * Returned Value:
+ *   None.
+ *
+ ****************************************************************************/
+
+void up_send_smp_call(cpu_set_t cpuset)
+{
+  up_trigger_irq(GIC_SMP_CALL, cpuset);
 }
 
 #endif /* CONFIG_SMP */
