+++ arch/arm/src/armv7-r/Make.defs
@@ -59,6 +59,6 @@ endif
 
 ifeq ($(CONFIG_SMP),y)
   CMN_ASRCS += arm_cpuhead.S
-  CMN_CSRCS += arm_cpustart.c arm_cpupause.c
+  CMN_CSRCS += arm_cpustart.c arm_smpcall.c
   CMN_CSRCS += arm_cpuidlestack.c arm_scu.c
 endif
