+++ arch/arm/src/cxd56xx/CMakeLists.txt
@@ -42,7 +42,7 @@ set(SRCS
 if(CONFIG_SMP)
   list(APPEND SRCS cxd56_cpuidlestack.c)
   list(APPEND SRCS cxd56_cpuindex.c)
-  list(APPEND SRCS cxd56_cpupause.c)
+  list(APPEND SRCS cxd56_smpcall.c)
   list(APPEND SRCS cxd56_cpustart.c)
   if(CONFIG_CXD56_TESTSET)
     list(APPEND SRCS cxd56_testset.c)
