+++ arch/arm/src/cxd56xx/Make.defs
@@ -41,7 +41,7 @@ CHIP_CSRCS += cxd56_sysctl.c
 ifeq ($(CONFIG_SMP), y)
 CHIP_CSRCS += cxd56_cpuidlestack.c
 CHIP_CSRCS += cxd56_cpuindex.c
-CHIP_CSRCS += cxd56_cpupause.c
+CHIP_CSRCS += cxd56_smpcall.c
 CHIP_CSRCS += cxd56_cpustart.c
 ifeq ($(CONFIG_CXD56_TESTSET),y)
 CHIP_CSRCS += cxd56_testset.c
