+++ arch/arm/src/cxd56xx/cxd56_cpustart.c
@@ -72,7 +72,7 @@
 
 volatile static spinlock_t g_appdsp_boot;
 
-extern int arm_pause_handler(int irq, void *c, void *arg);
+extern int cxd56_smp_call_handler(int irq, void *c, void *arg);
 
 /****************************************************************************
  * Private Functions
@@ -117,8 +117,8 @@ static void appdsp_boot(void)
 
   /* Enable SW_INT */
 
-  irq_attach(CXD56_IRQ_SW_INT, arm_pause_handler, NULL);
-  up_enable_irq(CXD56_IRQ_SW_INT);
+  irq_attach(CXD56_IRQ_SMP_CALL, cxd56_smp_call_handler, NULL);
+  up_enable_irq(CXD56_IRQ_SMP_CALL);
 
   spin_unlock(&g_appdsp_boot);
 
@@ -232,8 +232,8 @@ int up_cpu_start(int cpu)
 
       /* Setup SW_INT for this APP_DSP0 */
 
-      irq_attach(CXD56_IRQ_SW_INT, arm_pause_handler, NULL);
-      up_enable_irq(CXD56_IRQ_SW_INT);
+      irq_attach(CXD56_IRQ_SMP_CALL, cxd56_smp_call_handler, NULL);
+      up_enable_irq(CXD56_IRQ_SMP_CALL);
     }
 
   spin_lock(&g_appdsp_boot);
