+++ arch/arm/src/lc823450/lc823450_cpustart.c
@@ -69,7 +69,7 @@ static volatile spinlock_t g_cpu_wait[CONFIG_SMP_NCPUS];
  * Private Functions
  ****************************************************************************/
 
-extern int lc823450_pause_handler(int irq, void *c, void *arg);
+extern int lc823450_smp_call_handler(int irq, void *c, void *arg);
 
 /****************************************************************************
  * Name: cpu1_boot
@@ -106,8 +106,8 @@ static void cpu1_boot(void)
       up_enable_irq(LC823450_IRQ_MEMFAULT);
 #endif
 
-      irq_attach(LC823450_IRQ_CTXM3_01, lc823450_pause_handler, NULL);
-      up_enable_irq(LC823450_IRQ_CTXM3_01);
+      irq_attach(LC823450_IRQ_SMP_CALL_01, lc823450_smp_call_handler, NULL);
+      up_enable_irq(LC823450_IRQ_SMP_CALL_01);
     }
 
   spin_unlock(&g_cpu_wait[0]);
@@ -193,8 +193,8 @@ int up_cpu_start(int cpu)
 
   /* IRQ setup CPU1->CPU0 */
 
-  irq_attach(LC823450_IRQ_CTXM3_11, lc823450_pause_handler, NULL);
-  up_enable_irq(LC823450_IRQ_CTXM3_11);
+  irq_attach(LC823450_IRQ_SMP_CALL_11, lc823450_smp_call_handler, NULL);
+  up_enable_irq(LC823450_IRQ_SMP_CALL_11);
 
   spin_lock(&g_cpu_wait[0]);
