+++ arch/arm/src/rp2040/Make.defs
@@ -37,7 +37,7 @@ CHIP_CSRCS += rp2040_pll.c
 ifeq ($(CONFIG_SMP),y)
 CHIP_CSRCS += rp2040_cpuindex.c
 CHIP_CSRCS += rp2040_cpustart.c
-CHIP_CSRCS += rp2040_cpupause.c
+CHIP_CSRCS += rp2040_smpcall.c
 CHIP_CSRCS += rp2040_cpuidlestack.c
 CHIP_CSRCS += rp2040_testset.c
 CMN_ASRCS  := $(filter-out arm_testset.S,$(CMN_ASRCS))
