+++ arch/arm/src/sam34/Make.defs
@@ -138,7 +138,7 @@ endif # CONFIG_SAM34_TC
 
 ifeq ($(CONFIG_SMP),y)
 CHIP_CSRCS += sam4cm_cpuindex.c sam4cm_cpuidlestack.c
-CHIP_CSRCS += sam4cm_cpupause.c sam4cm_cpustart.c
+CHIP_CSRCS += sam4cm_smpcall.c sam4cm_cpustart.c
 ifneq ($(CONFIG_ARCH_IDLE_CUSTOM),y)
 CHIP_CSRCS += sam4cm_idle.c
 endif
