+++ arch/arm64/src/common/Make.defs
@@ -84,7 +84,7 @@ endif
 
 ifeq ($(CONFIG_SMP),y)
 CMN_CSRCS += arm64_cpuidlestack.c arm64_cpustart.c
-CMN_CSRCS += arm64_cpupause.c
+CMN_CSRCS += arm64_smpcall.c
 endif
 
 ifeq ($(CONFIG_BUILD_KERNEL),y)
