+++ arch/arm64/src/common/arm64_gic.h
@@ -281,15 +281,13 @@
 #define GIC_IRQ_SGI15               15
 
 #ifdef CONFIG_ARCH_TRUSTZONE_SECURE
-#  define GIC_SMP_CPUPAUSE_ASYNC    GIC_IRQ_SGI8
-#  define GIC_SMP_CPUSTART          GIC_IRQ_SGI9
-#  define GIC_SMP_CPUPAUSE          GIC_IRQ_SGI10
-#  define GIC_SMP_CPUCALL           GIC_IRQ_SGI11
+#  define GIC_SMP_SCHED             GIC_IRQ_SGI9
+#  define GIC_SMP_CPUSTART          GIC_IRQ_SGI10
+#  define GIC_SMP_CALL              GIC_IRQ_SGI11
 #else
-#  define GIC_SMP_CPUPAUSE_ASYNC    GIC_IRQ_SGI0
-#  define GIC_SMP_CPUSTART          GIC_IRQ_SGI1
-#  define GIC_SMP_CPUPAUSE          GIC_IRQ_SGI2
-#  define GIC_SMP_CPUCALL           GIC_IRQ_SGI3
+#  define GIC_SMP_SCHED             GIC_IRQ_SGI1
+#  define GIC_SMP_CPUSTART          GIC_IRQ_SGI2
+#  define GIC_SMP_CALL              GIC_IRQ_SGI3
 #endif
 
 /****************************************************************************
@@ -329,10 +327,10 @@ int arm64_gic_v2m_initialize(void);
 #ifdef CONFIG_SMP
 
 /****************************************************************************
- * Name: arm64_pause_async_handler
+ * Name: arm64_smp_sched_handler
  *
  * Description:
- *   This is the handler for async pause.
+ *   This is the handler for sched.
  *
  *   1. It saves the current task state at the head of the current assigned
  *      task list.
@@ -348,7 +346,7 @@ int arm64_gic_v2m_initialize(void);
  *
  ****************************************************************************/
 
-int arm64_pause_async_handler(int irq, void *context, void *arg);
+int arm64_smp_sched_handler(int irq, void *context, void *arg);
 
 void arm64_gic_secondary_init(void);
