+++ arch/risc-v/src/common/CMakeLists.txt
@@ -44,7 +44,7 @@ if(NOT CONFIG_ALARM_ARCH)
 endif()
 
 if(CONFIG_SMP)
-  list(APPEND SRCS riscv_cpuindex.c riscv_cpupause.c riscv_cpustart.c)
+  list(APPEND SRCS riscv_cpuindex.c riscv_smpcall.c riscv_cpustart.c)
 endif()
 
 if(CONFIG_RISCV_MISALIGNED_HANDLER)
