+++ arch/risc-v/src/common/riscv_exception.c
@@ -298,7 +298,7 @@ void riscv_exception_attach(void)
   irq_attach(RISCV_IRQ_RESERVED14, riscv_exception, NULL);
 
 #ifdef CONFIG_SMP
-  irq_attach(RISCV_IRQ_SOFT, riscv_pause_handler, NULL);
+  irq_attach(RISCV_IRQ_SOFT, riscv_smp_call_handler, NULL);
 #else
   irq_attach(RISCV_IRQ_SOFT, riscv_exception, NULL);
 #endif
