+++ arch/sparc/src/s698pm/s698pm-irq.c
@@ -185,7 +185,7 @@ int s698pm_cpuint_initialize(void)
 #if defined CONFIG_SMP
   /* Attach IPI interrupts */
 
-  irq_attach(S698PM_IPI_IRQ, s698pm_pause_handler, NULL);
+  irq_attach(S698PM_IPI_IRQ, s698pm_smp_call_handler, NULL);
 
   (void)s698pm_setup_irq(cpu, S698PM_IPI_IRQ, 0);
