+++ arch/x86_64/src/intel64/CMakeLists.txt
@@ -61,7 +61,7 @@ if(CONFIG_ARCH_HAVE_TESTSET)
 endif()
 
 if(CONFIG_SMP)
-  list(APPEND SRCS intel64_cpuidlestack.c intel64_cpupause.c intel64_cpustart.c)
+  list(APPEND SRCS intel64_cpuidlestack.c intel64_smpcall.c intel64_cpustart.c)
 endif()
 
 if(CONFIG_MULTBOOT2_FB_TERM)
