+++ arch/x86_64/src/intel64/Make.defs
@@ -48,7 +48,7 @@ endif
 
 ifeq ($(CONFIG_SMP),y)
 CHIP_CSRCS += intel64_cpuidlestack.c
-CHIP_CSRCS += intel64_cpupause.c
+CHIP_CSRCS += intel64_smpcall.c
 CHIP_CSRCS += intel64_cpustart.c
 endif
