+++ arch/x86_64/src/intel64/intel64_cpustart.c
@@ -47,8 +47,8 @@
  ****************************************************************************/
 
 extern void __ap_entry(void);
-extern int up_pause_handler(int irq, void *c, void *arg);
-extern int up_pause_async_handler(int irq, void *c, void *arg);
+extern int x86_64_smp_call_handler(int irq, void *c, void *arg);
+extern int x86_64_smp_sched_handler(int irq, void *c, void *arg);
 
 /****************************************************************************
  * Private Functions
@@ -164,10 +164,10 @@ void x86_64_ap_boot(void)
 
   /* Connect Pause IRQ to CPU */
 
-  irq_attach(SMP_IPI_IRQ, up_pause_handler, NULL);
-  irq_attach(SMP_IPI_ASYNC_IRQ, up_pause_async_handler, NULL);
-  up_enable_irq(SMP_IPI_IRQ);
-  up_enable_irq(SMP_IPI_ASYNC_IRQ);
+  irq_attach(SMP_IPI_CALL_IRQ, x86_64_smp_call_handler, NULL);
+  irq_attach(SMP_IPI_SCHED_IRQ, x86_64_smp_sched_handler, NULL);
+  up_enable_irq(SMP_IPI_CALL_IRQ);
+  up_enable_irq(SMP_IPI_SCHED_IRQ);
 
 #ifdef CONFIG_STACK_COLORATION
   /* If stack debug is enabled, then fill the stack with a
