+++ arch/xtensa/src/esp32s3/esp32s3_intercpu_interrupt.c
@@ -64,9 +64,9 @@ static int IRAM_ATTR esp32s3_fromcpu_interrupt(int irq, void *context,
                              SYSTEM_CPU_INTR_FROM_CPU_1_REG;
   putreg32(0, regaddr);
 
-  /* Call pause handler */
+  /* Smp call handler */
 
-  xtensa_pause_handler(irq, context, arg);
+  xtensa_smp_call_handler(irq, context, arg);
 
   return OK;
 }
